datalog template: wip
