+++ src/Language/DifferentialDatalog/rust_template/lib.rs
@@ -1,0 +1,16 @@
+#[macro_use]
+extern crate abomonation;
+extern crate num;
+
+#[macro_use] 
+extern crate serde_derive;
+extern crate serde;
+
+extern crate timely;
+extern crate timely_communication;
+extern crate differential_dataflow;
+
+mod uint;
+mod variable;
+pub mod program;
+mod test;
