+++ src/Language/DifferentialDatalog/rust_template/variable.rs
@@ -1,0 +1,53 @@
+use timely::progress::nested::product::Product;
+use timely::dataflow::*;
+use timely::dataflow::scopes::{Child};
+use timely::dataflow::operators::*;
+use timely::dataflow::operators::feedback::Handle;
+use differential_dataflow::{Data, Collection, Hashable};
+use differential_dataflow::operators::*;
+use differential_dataflow::lattice::Lattice;
+
+/// A collection defined by multiple mutually recursive rules.
+///
+/// A `Variable` names a collection that may be used in mutually recursive rules. This implementation
+/// is like the `Variable` defined in `iterate.rs` optimized for Datalog rules: it supports repeated
+/// addition of collections, and a final `distinct` operator applied before connecting the definition.
+pub struct Variable<'a, G: Scope, D: Default+Data+Hashable>
+where G::Timestamp: Lattice+Ord {
+    feedback: Option<Handle<G::Timestamp, u64,(D, Product<G::Timestamp, u64>, isize)>>,
+    current: Collection<Child<'a, G, u64>, D>,
+    cycle: Collection<Child<'a, G, u64>, D>,
+}
+
+impl<'a, G: Scope, D: Default+Data+Hashable> Variable<'a, G, D> where G::Timestamp: Lattice+Ord {
+    /// Creates a new `Variable` from a supplied `source` stream.
+    pub fn from(source: &Collection<Child<'a, G, u64>, D>) -> Variable<'a, G, D> {
+        let (feedback, cycle) = source.inner.scope().loop_variable(u64::max_value(), 1);
+        let cycle = Collection::new(cycle);
+        let mut result = Variable { feedback: Some(feedback), current: cycle.clone(), cycle: cycle };
+        result.add(source);
+        result
+    }
+    /// Adds a new source of data to the `Variable`.
+    pub fn add(&mut self, source: &Collection<Child<'a, G, u64>, D>) {
+        self.current = self.current.concat(source);
+    }
+}
+
+impl<'a, G: Scope, D: Default+Data+Hashable> ::std::ops::Deref for Variable<'a, G, D> where G::Timestamp: Lattice+Ord {
+    type Target = Collection<Child<'a, G, u64>, D>;
+    fn deref(&self) -> &Self::Target {
+        &self.cycle
+    }
+}
+
+impl<'a, G: Scope, D: Default+Data+Hashable> Drop for Variable<'a, G, D> where G::Timestamp: Lattice+Ord {
+    fn drop(&mut self) {
+        if let Some(feedback) = self.feedback.take() {
+            self.current.distinct()
+                        .inner
+                        .connect_loop(feedback);
+        }
+    }
+}
+
